Add new modules for helpers and form
- Add default location to display when user opens the website
