+++ src/assets/stylesheets/style.scss
@@ -37,3 +37,7 @@ article {
     }
   }
 }
+
+.error-msg {
+  display: none;
+}
